Move classproperty decorator into utils
